+++ lua/cmp_obsidian_new.lua
@@ -41,7 +41,11 @@ source.complete = function(self, request, callback)
     local new_title, new_id, path, rel_path
     new_id = client:new_note_id(search)
     new_title, new_id, path = client:parse_title_id_path(search, new_id, dir)
-    rel_path = assert(client:vault_relative_path(path))
+    rel_path = client:vault_relative_path(path)
+    if rel_path == nil then
+      log.err("Failed to resolve path '%s' relative to vault root '%s'", path, client:vault_root())
+      return
+    end
     if vim.endswith(rel_path, ".md") then
       rel_path = string.sub(rel_path, 1, -4)
     end